requirements.txt
Flask==2.3.3
Pillow==10.2.0
qrcode==7.3.1
gunicorn==21.2.0
google-auth==2.20.0
google-api-python-client==2.88.0